Adds Class TopologyMap to the project.
The Class TopologyMap represents a geopositional map for the application and
helps to decide when a Join at Back could be started.

This class is still in early stages of development and could permit/deny
vehicles to join at the back of a platoon. We envisionate this class to
serve as a powerful decision mechanism to suggest maneuvers based on current
platoon topology.